--- src/main/java/roomescape/admin/AdminController.java
@@ -1,13 +1,0 @@
-package roomescape.admin;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class AdminController {
-    
-    @GetMapping("/admin")
-    public String adminPage() {
-        return "admin/index";
-    }
-}
