+++ src/main/java/roomescape/reservation/controller/ReservationAdminPageController.java
@@ -1,0 +1,28 @@
+package roomescape.reservation.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ReservationAdminPageController {
+
+    @GetMapping("/admin")
+    public String adminPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String adminReservationPage() {
+        return "admin/reservation-new";
+    }
+
+    @GetMapping("/admin/time")
+    public String timePage() {
+        return "admin/time";
+    }
+
+    @GetMapping("/admin/theme")
+    public String adminThemePage() {
+        return "admin/theme";
+    }
+}
