--- src/main/java/roomescape/reservation/controller/ThemePageController.java
@@ -1,18 +1,0 @@
-package roomescape.reservation.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class ThemePageController {
-
-    @GetMapping("/admin/theme")
-    public String adminThemePage() {
-        return "admin/theme";
-    }
-
-    @GetMapping("/")
-    public String popularPage() {
-        return "index";
-    }
-}
